test 브렌치 작업 시작

+++ src/main/java/ac/su/kiosk/repository/MenuRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MenuRepository extends JpaRepository<Menu, Long> {
+public interface MenuRepository extends JpaRepository<Menu, Integer> {
     List<Menu> findAllByCategory(Category category);
 }
